requirements.txt
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
einops==0.8.0
fasttext-wheel==0.9.2
filelock==3.16.1
flashtext==2.7
fonttools==4.54.1
fsspec==2024.10.0
gensim==4.3.3
hazm==0.10.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
lightning-utilities==0.11.8
MarkupSafe==3.0.2
matplotlib==3.9.2
mpmath==1.3.0
networkx==3.4.2
nltk==3.9.1
numpy==1.24.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.3
pillow==11.0.0
pybind11==2.13.6
pyparsing==3.2.0
python-crfsuite==0.9.11
python-dateutil==2.9.0.post0
pytz==2024.2
regex==2024.9.11
scikit-learn==1.5.2
scipy==1.13.1
six==1.16.0
smart-open==7.0.5
sympy==1.13.1
threadpoolctl==3.5.0
torch==2.5.0
torchmetrics==1.5.1
torchvision==0.20.0
tqdm==4.66.5
triton==3.1.0
typing_extensions==4.12.2
tzdata==2024.2
wrapt==1.16.0